Combine `hotkey` and `normalize-hotkey` files and add a branded type
